Create function for removing a task
The function for removing a task lives in App. It uses setState to filter the tasks based on the id of the task clicked. The id is also the key (uniqid), duplicated.

The onTaskRemoval is passed as a prop to the Overview so that tasks can call it when clicked.