Add edit functionality to each task
The core methods are inside the App. Rendering the Overview now has a conditional based on a new property of each task: edit. If edit is false, the usual components are rendered. If edit is true, components are rendered for editing.

The two new methods (onTaskEdit and onTaskSubmit) use map to alter the state of the tasks. Of course once that alteration takes place, the Overview re-renders accordingly.